refactor: добавлена типизация формы входа
