+++ client/src/pages/LoginPage/loginFormTypes.ts
@@ -1,0 +1,8 @@
+export const LoginFormFields = {
+    login: "login",
+    password: "password"
+} as const;
+
+export type LoginFormFields = typeof LoginFormFields[keyof typeof LoginFormFields];
+
+export type LoginFormState = Record<LoginFormFields, string>;
